changed quiz and owl, updated concepts.gam, other fixes
I deleted the old quiz, the owl quiz was moved to quiz.xhtml. I added
concepts as 'done' and 'knowledge' per subject in the concepts.gam.  The
list of courses is now in order on alphabet and at last the O.W.L. Exam
(seems more logical to me). Furthermore I added a dot and capital in the
cookie bar and explained something in the main school area.