Implementing part of the OWL test
What is does: lets you choose options per subject and evaluate the test.
At the top, you can see your point (0, 50 or 100). No additional
comments are given, because it then opens in a new html page. You can do
the test over and over. It DOESN'T process failure of a course to the
knowledge level of that course, and I would like the test part to
disappear when the a person scored 100, but couldn't figure out how I
can evaluate a value of a concept in a statement...